Split main and functions in two different files
